backend/strategy_runner/requirements.txt
# Strategy Runner Dependencies

# Core dependencies
pandas==2.0.3
pandas-ta==0.3.14b0
numpy==1.24.3

# Async support
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging and monitoring
structlog>=23.0.0

# Time and timezone handling
python-dateutil>=2.8.0
pytz>=2023.3

# HTTP client
requests>=2.31.0

# Data validation
marshmallow>=3.20.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

